Use assembly for MMIO reads and writes on aarch64.
volatile_read and volatile_write doesn't work properly in a protected
VM.
